cleaned up and moved Google Calander to its own service
