+++ backend/api/MirrorOfErised/MirrorOfErised.models/MirrorOfErised.models.csproj
@@ -22,6 +22,7 @@
 
   <ItemGroup>
     <Folder Include="Migrations" />
+    <Folder Include="Services" />
   </ItemGroup>
 
 </Project>
